+++ Source/WebCore/platform/graphics/cocoa/SystemFontDatabaseCoreText.cpp
@@ -29,6 +29,7 @@
 #include "FontCache.h"
 #include "FontCacheCoreText.h"
 #include "FontCascadeDescription.h"
+#include "FontMetricsNormalization.h"
 
 #include <wtf/cf/TypeCastsCF.h>
